Extract WeatherNow component from Main
