+++ src/app/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import weatherNowReducer from "../features/weatherNow/weatherNowSlice";
+import cityListReducer from "../features/cityList/cityListSlice";
 
 export default configureStore({
   reducer: {
-    weatherNow: weatherNowReducer
+    weatherNow: weatherNowReducer,
+    cityList: cityListReducer
   }
 });
